Debounce input events to once per 100ms
This drastically reduces the number of requests sent to the backend
when typing quickly. Individual requests return with more than
acceptable performance (~150ms), but if too many are sent too quickly,
they tend to pile up to almost 5000ms!

The tradeoff here is a slightly increased delay for autocomplete. I
tried 500ms down to 50ms and this seems like a good middle ground.